foomatic-rip: Corrected declaration of print_pdf() function in pdf.h

+++ filter/foomatic-rip/pdf.h
@@ -24,7 +24,7 @@
 #ifndef pdf_h
 #define pdf_h
 
-int print_pdf(FILE *s, const char *alreadyread, size_t len, const char *filename, int startpos);
+int print_pdf(FILE *s, const char *alreadyread, size_t len, const char *filename, size_t startpos);
 
 #endif
